+++ CHANGELOG.md
@@ -22,6 +22,8 @@ All notable changes to this project will be documented in this file.
  - Every element now have a `visible` property
  - Added `changed` callback to the `Slider` widget
  - Added `TabWidget` widget
+ - Rust: `sixtyfps::Image` can now be constructed from image data provided by `sixtyfps::SharedImageBuffer`.
+   This enables integrating with other low-level software rendering or the the popular Rust image crate.
 
 ### Fixed